skills/shared-scripts/pyproject.toml
[project]
name = "b2b-shared"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "brightdata-sdk>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
